A5b: adjudicate A5a (Fable) - agree 13, partly agree 4, refute 0
Independent review first, then item-by-item rulings on A5a's 17 findings.
All four high-severity items stand, none were refuted.

#3 axis bug CONFIRMED and fixed. Beyond the synthetic check, A5b tested it
against real data: across 1,233 train+val masks the H-axis lateral fraction
concentrates at 0 or 1 for 79.7% of subjects, as a lateralised lesion
requires, versus 50.4% for W. features.py is corrected; dev LAA-vs-CE rises
0.6551 -> 0.6709 [0.611, 0.730] with macro unchanged.

#2 test-reading CLAUDE.md rule rewritten to be true rather than
convenient. Building a stratified split requires whole-cohort mask volumes,
so build_index.py is now an explicit, narrowly scoped exception; A2's
a2_volumes.py full-cohort re-read is recorded by name and timestamp, now
defaults to train+val, and an allow-list regression test enforces the rule.

#1 charter v2 relabelled "revised after dev analysis, before test
evaluation", and the v1 criterion will be reported alongside on test.

#4 partly agreed: seg-train predicted masks are measurably cleaner
(Dice 0.758 vs 0.706) but splitting pooled OOF AUC shows no detectable
bias (full macro 0.717 vs 0.720), so the dev experiments stand with the
non-independence documented.

No retraining. The flip is refactored into an explicit flip_axis option,
and flip_axis: ap reproduces best.pt byte-identically (e10d9268...), so the
existing run stays valid. A 7-run, 3-seed comparison shows the true
left-right flip gains Dice (0.6909 +-0.0024 vs 0.6817 +-0.0035, same sign
on every seed) but loses volume ICC (0.850 vs 0.914), so it is not a clean
win, and adopting it would spend another val selection after the criteria
were frozen. Documented as a non-recommended alternative instead.

Recorded honestly in charter section C: across the 7 runs val ICC ranges
0.824-0.914 and the frozen checkpoint holds the maximum, so S3 (ICC >= 0.85)
may well fail on test.

Test evaluation: CONDITIONALLY APPROVED. Conditions include reporting both
detection rates (non-empty output 0.9725 and lesion-contact 0.9312), CIs on
every AUC, E1 interpreted at its real 0.54 PASS-rule power rather than the
0.93 point-estimate power, and no "never read the test set" phrasing.

pytest: 46 -> 52 passed.

+++ scripts/analysis/a6_threshold.py
@@ -38,7 +38,8 @@ def load_probs(run: Path, cache_dir: Path, ids: list[str], ckpt: str, cache_npz:
     device = "cuda" if torch.cuda.is_available() else "cpu"
     model.to(device)
     cache = SubjectCache(cache_dir, ids)
-    probs = {sid: predict_subject(model, cache.img[sid], device=device, channels=channels).astype(np.float16) for sid in ids}
+    # fp32, not float16: a float16 round trip rounds e.g. 0.49999 up to 0.5 and flips the 0.5 decision (A5a item 15)
+    probs = {sid: predict_subject(model, cache.img[sid], device=device, channels=channels).astype(np.float32) for sid in ids}
     meta = {sid: {"gt_ml": volume_ml(cache.mask[sid] > 0, cache.meta[sid]["voxel_volume_mm3"]),
                   "vv": float(cache.meta[sid]["voxel_volume_mm3"])} for sid in ids}
     gts = {sid: (cache.mask[sid] > 0) for sid in ids}
